+++ app/src/main/java/com/dskja/betterstreamflix/fragments/settings/SettingsTvFragment.kt
@@ -96,7 +96,7 @@ class SettingsTvFragment : LeanbackPreferenceFragmentCompat() {
     )
 
     private val DEFAULT_DOMAIN_VALUE = "streamingunity.win"
-    private val DEFAULT_SERIENSTREAM_DOMAIN_VALUE = "serienstream.to"
+    private val DEFAULT_SERIENSTREAM_DOMAIN_VALUE = "186.2.175.5"
     private val DEFAULT_MOFLIX_DOMAIN_VALUE = "moflix-stream.xyz"
     private val DEFAULT_CUEVANA_DOMAIN_VALUE = "cuevana3.gs"
     private val DEFAULT_POSEIDON_DOMAIN_VALUE = "www.poseidonhd2.co"